Fixed gcc printf-format warnings.

+++ src/sbbs3/logon.cpp
@@ -126,9 +126,9 @@ bool sbbs_t::logon()
 			,cfg.cdt_per_dollar*cfg.node_dollars_per_call);
 		bprintf(text[CreditedAccount]
 			,cfg.cdt_per_dollar*cfg.node_dollars_per_call);
-		sprintf(str,"%s #%u was billed $%d T: %u seconds"
+		sprintf(str,"%s #%u was billed $%d T: %lu seconds"
 			,useron.alias,useron.number
-			,cfg.node_dollars_per_call,(uint)now-answertime);
+			,cfg.node_dollars_per_call,now-answertime);
 		logline("$+",str);
 		hangup();
 		return(false); }
